test: :white_check_mark: fix all benchmark functions to use the new random tensors generator APIs
In the previous commit, I replaced the `generateRandomArrays` API with two new APIs, `generateRandomTensor2DSlice` and `generateRandomTensor3DSlice`, to support the generation of slices of random 2D tensors and 3D tensors for testing and benchmarking.

I updated the API usage in the test functions but mistakenly forgot to update the API usage in the benchmark functions, causing the CI tests to fail.

This patch aims to update the benchmark functions to use the new APIs and get the CI tests passing again.