refactor!: :recycle: reorganize source files to improve the project structure
* I've moved source files I do not want to expose to client applications into the `internal` folder.
* I've introduced separate folders for black-box `tests` and `benchmarks` into the project structure.

BREAKING CHANGE: I've replaced the `Relation` type with the `int` type to indicate whether a value is greater than, equal to, or lesser than another value. I made this change to make the APIs more consistent with the practices followed in the Go standard library's APIs (e.g. `slices.SortFunc`) and to break import cycles.