refactor!: :recycle: replace `Add` API with `Insert` API
I have replaced the `Add` API with a new `Insert` API. I made this change to make the APIs of my K-D Tree library consistent with those of existing ones written in Go.

BREAKING CHANGE: Other than the name change, the new `Insert` API does not return any value, unlike the previous `Add` API, which would return a boolean value indicating whether the element was inserted into the K-D Tree or skipped because it was a duplicate. With the removal of the `Add` API in this commit, any code that uses the `Add` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the `Insert` API to be compatible with future releases of this library.