perf!: :zap: improve the time and space efficiency of the `KNearestNeighbor` API
The fundamental changes made to improve the time and space efficiency of the `KNearestNeighbor` API are:
1. Replace the interface-based implementation of the priority queue with a generics-based implementation.
2. Preallocate an underlying slice with a capacity equal to the maximum allowed size of the priority queue.
3. The `Item` struct now stores a pointer to the original data instead of storing a copy of the original data.
4. Return a slice of pointers to the k-nearest neighbours (i.e., `[]*T`) instead of a slice of copies of the k-nearest neighbours' values (i.e., `[]T`).

These changes have led to a remarkable ~34.26% enhancement in performance (i.e., `ns/op`). The memory efficiency has also substantially improved, with ~5 times less `B/op` and ~113 times less `allocs/op`.

BREAKING CHANGE: To improve the performance of the `KNearestNeighbor` API, I changed its return type from `[]T` to `[]*T` to reduce the number of costly data copies that occur. Utilizing pointers to access the actual data is a significantly more efficient approach in terms of CPU and memory usage. With just a single level of indirection, we can save on a considerable amount of data copying, ensuring a more streamlined implementation. With this change to the `KNearestNeighbor` API's return type in this commit, any code that uses the `KNearestNeighbor` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the updated version of the `KNearestNeighbor` API to be compatible with future releases of this library.