refactor!: :recycle: rename `KNearestNeighbor` API to the `KNN` API
I have renamed the `KNearestNeighbor` API to the `KNN` API. I made this change to make the APIs of my K-D Tree library consistent with those of existing ones written in Go.

BREAKING CHANGE: With the rename of the `KNearestNeighbor` API to the `KNN` API in this commit, any code that uses the `KNearestNeighbor` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the `KNN` API to be compatible with future releases of this library.