refactor!: :recycle: rename `Query` API to the `RangeSearch` API
I have renamed the `Query` API to the `RangeSearch` API. I made this change to make the APIs of my K-D Tree library consistent with those of existing ones written in Go.

BREAKING CHANGE: This commit renames the `Query` API to the `RangeSearch` API. Any code that uses the `Query` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the `RangeSearch` API to be compatible with future releases of this library.