feat: :sparkles: add an API to visualize the K-D Tree using Graphviz
I have introduced a new `Dot` API, which returns a string containing DOT (graph description language) code. The user can use this DOT code to visualize the current state of their given K-D Tree using Graphviz, providing a clear and comprehensive view of its structure and contents.

The ability to visualize the current state of the K-D Tree helps students and newcomers understand how it works.

We can also use the `Dot` API to debug the K-D Tree. It provides a clear representation of its current state, making it easier to identify and fix any issues, which is beneficial for both the library author and its users.

Note that by "state" of the K-D Tree, we refer to its structure and constituent nodes' contents. The contents of a node are obtained from the output of calling the `String` API on its constituent value.

I was inspired to write this API after reading a blog post from Eli Bendersky.
- Link to the blog post: https://eli.thegreenplace.net/2009/11/23/visualizing-binary-trees-with-graphviz