refactor!: :recycle: rename `Delete` API to the `Remove` API
I have renamed the `Delete` API to the `Remove` API. I made this change to make the APIs of my K-D Tree library consistent with those of existing ones written in Go.

BREAKING CHANGE: With the rename of the `Delete` API to the `Remove` API in this commit, any code that uses the `Delete` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the `Remove` API to be compatible with future releases of this library.