refactor!: :recycle: replace the `KNN` API's return type from `[]*T` to `[]T`
I have replaced the `KNN` API's return type from `[]*T` to `[]T`. I made this change to make the APIs of my K-D Tree library consistent with those of existing ones written in Go.

BREAKING CHANGE: The updated `KNN` API returns a `[]T`, i.e., slice of elements of type `T`, unlike the previous `KNN` API, which would return a `[]*T`, i.e., slice of pointers to elements of type `T`. With this change in the `KNN` API in this commit, any code that uses the `KNN` API will no longer work with future releases of this library created after this commit. Affected codebases must migrate to the updated `KNN` API to be compatible with future releases of this library.