test: :white_check_mark: add benchmarking setup to compare against `kyroy/kdtree`
The benchmarking setup I've added compares the API performance between my library and the `kyroy/kdtree` library, providing valuable insights into the performance characteristics of each one.

The APIs compared against each other are listed in the table below:

| S. No. | rishitc/go-kd-tree  | kyroy/kdtree |
|:------:|---------------------|--------------|
|    1   | NewKDTreeWithValues | New          |
|    2   | RangeSearch         | RangeSearch  |
|    3   | Values              | Points       |
|    4   | Insert              | Insert       |
|    5   | Remove              | Remove       |
|    6   | Balance             | Balance      |
|    7   | KNN                 | KNN          |

Certain APIs, such as `String,` have not been benchmarked because they are not performance-critical and are primarily used for debugging and educational purposes. Additionally, there is no standard way of representing a tree as a string, so the benchmark results are heavily dependent on your library's chosen format of data representation as a string.

In this benchmarking setup, we no longer generate the benchmarking dataset on the fly every time we benchmark. Instead, we generate the benchmarking dataset once and store it in a `.csv` file (if it does not already exist), which is later read and stored in memory before running any benchmarks.
This new approach significantly reduces the time taken to run a benchmark.

To simplify running the benchmarking setup, which uses build flags, I've included a template command in the file `internal/benchmarks/README.md` that can be easily modified and used to run the benchmarks locally.

Finally, I have removed the file used to benchmark old and new approaches for implementing different APIs.